\name{inla.models}
\alias{inla.models}

\title{Valid models in INLA} 
\description{This page describe the models implemented in inla, divided into sections: latent, group, mix, link, predictor, hazard, likelihood, prior, wrapper.}
\usage{inla.models()}
\value{see \code{\link[INLA]{inla.models}}}
\details{
In order for the INLA package to be in Import rather than Depend, the inla.models
function must be exported.  Hence the need for this documentation.
}
