\name{inla}
\alias{inla}
\alias{inla.models}
\alias{inla.make.lincombs}

\title{
Functions from INLA
}
\description{
Some functions from INLA which need documentation in geostatsinla
}
\usage{
inla.models(...)
}
\arguments{
  \item{\dots}{arguments passed to \code{\link[INLA]{INLA}}}
}

\details{
In an effort to have the geostatsinla package build on r-forge, 
the INLA package is 'suggested' rather than 'Imports'.  As a result the 
functions here must be exported by geostatsinla and have documention.
}

 
\seealso{
\code{\link[INLA]{inla}}, \code{\link[INLA]{inla.models}}, 
\code{\link[INLA]{inla.make.lincombs}}, \code{\link[INLA]{inla.make.lincomb}}
}
 

